refac: move code in `array_to_vector.h` to `container_util.h`
