+++ tachyon/zk/plonk/examples/BUILD.bazel
@@ -57,7 +57,7 @@ tachyon_cc_library(
         ":simple_circuit_test_data",
         ":simple_lookup_circuit",
         ":simple_lookup_circuit_test_data",
-        "//tachyon/base:array_to_vector",
+        "//tachyon/base/containers:container_util",
         "//tachyon/math/elliptic_curves/bn/bn254",
         "//tachyon/zk/lookup:lookup_pair",
         "//tachyon/zk/plonk/examples/fibonacci:fibonacci1_circuit",
