+++ tachyon/zk/plonk/examples/circuit_test.h
@@ -7,7 +7,7 @@
 
 #include "absl/types/span.h"
 
-#include "tachyon/base/array_to_vector.h"
+#include "tachyon/base/containers/container_util.h"
 #include "tachyon/math/elliptic_curves/bn/bn254/bn254.h"
 #include "tachyon/zk/lookup/lookup_pair.h"
 #include "tachyon/zk/plonk/examples/circuit_test_type_traits.h"
